+++ packages/react-native-renderer/src/ReactNativeRenderer.js
@@ -25,7 +25,6 @@ import ReactNativeComponent from './ReactNativeComponent';
 import * as ReactNativeComponentTree from './ReactNativeComponentTree';
 import ReactNativeFiberRenderer from './ReactNativeFiberRenderer';
 import {getInspectorDataForViewTag} from './ReactNativeFiberInspector';
-import createReactNativeComponentClass from './createReactNativeComponentClass';
 import {injectFindHostInstance} from './findNodeHandle';
 import findNumericNodeHandle from './findNumericNodeHandle';
 
@@ -98,7 +97,6 @@ const ReactNativeRenderer: ReactNativeType = {
     NativeMethodsMixin,
     // Used by react-native-github/Libraries/ components
     ReactNativeComponentTree, // ScrollResponder
-    createReactNativeComponentClass, // RCTText, RCTView, ReactNativeART
     computeComponentStackForErrorReporting,
   },
 };
