--- packages/react-native-renderer/src/__mocks__/View.js
@@ -1,18 +1,0 @@
-/**
- * Copyright (c) 2013-present, Facebook, Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-'use strict';
-
-const createReactNativeComponentClass = require('../createReactNativeComponentClass')
-  .default;
-
-const View = createReactNativeComponentClass({
-  validAttributes: {},
-  uiViewClassName: 'View',
-});
-
-module.exports = View;
