+++ packages/react-native-renderer/src/__tests__/ReactNativeMount-test.internal.js
@@ -22,8 +22,8 @@ describe('ReactNative', () => {
     React = require('react');
     ReactNative = require('react-native-renderer');
     UIManager = require('UIManager');
-    createReactNativeComponentClass = require('../createReactNativeComponentClass')
-      .default;
+    createReactNativeComponentClass = require('ReactNativeViewConfigRegistry')
+      .register;
   });
 
   it('should be able to create and render a native component', () => {
