+++ scripts/rollup/bundles.js
@@ -124,7 +124,6 @@ const bundles = [
       'RCTEventEmitter',
       'TextInputState',
       'UIManager',
-      'View',
       'deepDiffer',
       'deepFreezeAndThrowOnMutationInDev',
       'flattenStyle',
@@ -147,7 +146,6 @@ const bundles = [
       'TextInputState',
       'UIManager',
       'FabricUIManager',
-      'View',
       'deepDiffer',
       'deepFreezeAndThrowOnMutationInDev',
       'flattenStyle',
